feat: add script letters
Getting script letters akin to `$\mathscr(P)$` is actually possible in Typst. It consists of changing the `stylistic-set` of the `text()` function to swap out the normal calligraphic set unicode range.

Which set that is, depends on the font used. In this case it's 1.
Adding this sentence would offset the footer and move it to a third page but I propose just removing the `unavailable` annotation as it is no longer needed after this and puts everything neatly back into two pages.